src/pvi-form-engine-release_unstructured_7.0/unstructured/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.98
botocore==1.34.98
cached_path==1.6.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
commonmark==0.9.1
confection==0.1.4
cymem==2.0.8
datasets==2.19.1
dill==0.3.8
distro==1.9.0
fairscale==0.4.6
fastcoref==2.1.6
filelock==3.8.2
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-cloud-translate==3.15.3
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
konoha==5.5.6
langcodes==3.4.0
language_data==1.2.0
Levenshtein==0.25.1
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
natto-py==1.0.1
negspacy==1.0.4
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.1.17
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
openai==1.36.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.6
redis==5.0.7
regex==2024.4.28
requests==2.31.0
rich==12.6.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.13.0
sentence-transformers==2.7.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.19.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0
tqdm==4.66.4
transformers==4.40.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
